Implement /NetworkDriver.DeleteEndpoint
This patch replaces the mocked version of /NetworkDriver.DeleteEndpoint
with the actual Neutron call. The unit test for the endpoint is also
implemented.

When the requst to delete the Endpoint is invoked, Kuryr deletes the
two or three resources corresponds to the Endpoint in the process of the
endpoint creation, i.e., Neutron port, Neutron subnet ofthe IPv4 address
and/or Neutron subnet of the IPv6 address. Kuryr searches the resources
with the given NetworkID and deletes them all based on the assumption
there's no conflict with the NetworkIDs.

Kuryr tries to delete all the Neutron subnets which IDs are contained in
the Neutron port, however if the multiple Docker endpoints have the same
subnet CIDR, the subntes can be referred from the multiple ports created
with the different Docker endpoint IDs and they can't be deleted unless
they're not referred from any port. In this case we'll get 409 Conflict
as the response status code and the corresponding exceptions. Kuryr
catches the exceptions, log them and continue to proceed because it is
totally the normal case.

The operation should be atomic but in this patch they're not because
there's no transaction mechanism in python-neutronclien and the better
design is required in the future.

Change-Id: I4c94428c613094e758bd228f586cb73f507b5775
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>