Move fake response generators to the base class
This patch moves _get_fake_subnets and _get_fake_ports from
TestKuryrEndpointFailures to TestKuryrBase to make it available for
TestKuryr. TestKuryr now uses them in the test cases to elimiate the
duplications in mocking the response from Neutron.

This patch also replaces calls of static methods with super because
they're not necessary actually and verbose.

Change-Id: I9bfc15e46fb626a9697137c7a8540ebfe8a6fa29
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>