Fix Kuryr to create the default subnets and return allocated CIDRs
Previously Kuryr queried the subnets with the names of the subnetpools
and therefore it always concluded there was no default subnet and
created a new one.

This patch fixes that behaviour and makes Kurry create new default
subnets associating them with the pre-defined subnetpools which names
would be given in the config file if they don't exit yet. If they exist,
Kuryr reuse them.

This patch also fixes Kuryr to return CIDRs combining the subnet CIDRs
with the IP addresses allocated by Neutorn's IPAM. This covers the case
none of Address and AddressIPv6 is given in the request.

Change-Id: If15e2509fdb28599cd5e9a0d21464bc97ea2dcb1
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>
Closes-Bug: #1496284