Add subnet mock staticmethods to TestKuryrBase for refactoring
This patch adds _get_fake_v4_subnet and _get_fake_v6_subnet to
TestKuryrBase and modifies TestKuryr to use them for mocking responses
for the subnet requests. Previously the mock data was hardcoded in the
test but they're used in different tests commonly and this patch
extracts them out not to lead the duplicated codes.

Change-Id: I7dc42d74e6246c7e4deb796c60deb2131731431d
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>