Modify Kuryr not to delete the default subnets
This patch fixes the bug that Kuryr deletes the default subnets
associated with the subnet pools for Kuryr specified in the config
file.

By default Kuryr creates the default subnets with the subnet pools. When
users delete the Docker endpoint, Kuryr deleted the default subnets.
However, the default subnets are used when users don't specify the CIDRs
for interfaces when they create Docker endpoints, so it's a better design
not to delete them.

Change-Id: I0b26895cba7003a4d2c6be4c22835f94ae6697d5
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>
Closes-Bug: #1497994