Fix /NetworkDriver.CreateEndpoint to handle the empty Interfaces
The current /NetworkDriver.CreateEndpiont doesn't handle the case
libnetwork gives the empty Interfaces property in the request and it's
failed when it receives the request which has the empty Interfaces
property. However, with the libnetwork CLI in Docker 1.8.0 we can't
specify the Interfaces property and it's always empty.

This patch makes Kuryr put a new entry with the ID 0 if it receives the
empty interface dictionary in the requests. MacAddress and Address or
AddressIPv6 are filled by allocated ones.

Change-Id: I689d2f3aa8c5d047dbdcfee425910951530fb621
Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>
Closes-Bug: #1499779