Create Neutron client from cloud credentials
This patch add the capabilities to create the neutron
client from admin credentials by using clouds.yaml.
This is done mostly for gate testing.

We will need to allow the user a configuration option to
disable this in production and use his own user/pass
for this authentication.

implements blueprint: fullstack-testing

Change-Id: I1973fa6fb13cedd60f86783165f949cb5b1e5f80