Using Neutron Tags
Uses Neutron tags for keeping track of the Docker network ID of the
corresponding Neutron network ID. Previously, the name attribute of
Neutron networks was used for holding the ID of the corresponding
Docker network. By using Neutron tags, we do not need to use the
Neutron network name for this purpose anymore.

As of now, there is a limit of 60 bytes for tags in Neutron that
prevents the use of 64-byte Docker IDs in one single tag. We therefore
use two tags.

The use of Neutron tags will allow a cleaner implementation of support
for existing Neutron network in a separate patchset.

Change-Id: I677876d0c4ff82e9d9b51cdd8d54633a66fd71ad