Support for existing Neutron nets
An existing Neutron network can be specified as the network
being used for a Docker network being created. The Neutron
network id or name can be specified.

If a Docker network is deleted, the corresponding pre existing
Neutron network will not be deleted.

Usage:
 docker network create -d kuryr --ipam-driver=kuryr
     --subnet=10.10.0.0/24 --gateway=10.10.0.1
     -o neutron.net.uuid=d98d1259-03d1-4b45-9b86-b039cba1d90d mynet

 docker network create -d kuryr --ipam-driver=kuryr
     --subnet=10.10.0.0/24 --gateway=10.10.0.1
     -o neutron.net.name=my_neutron_net mynet

If subnet being specified does not exist, it will be created by
Kuryr. If exists it will be used.

Blueprint existing-neutron-network

Change-Id: I570aa4e77c906d3f04768e882f02050cd0cad558