Adding support for two new libnetwork methods
In Docker 1.11 support for the two newly added
libnetwork methods is not required. However, not
having these supported result in errors that may
be confusing to users.

These two methods are to make support for exported
and published ports on a per container basis as discussed
here: https://github.com/docker/libnetwork/issues/992
It is proposed that by Docker version 1.12 support
for these methods become mandatory.

Support for exposed port will be added in a separate patchset.
Support for publishing port may be added later on when
needed support in Neutron is available.

Change-Id: Idf69a094eda2141cfa25b79d3095010871e1c9e8