startup: Fix Kuryr initialization
While testing Kuryr, I noticed it was not correctly detecting that the "tag"
service plugin was loaded. After some investigation with Mohammand, he
noticed that commit 7f4146e044f371c00825fb838a9cde5d173a92fe changed the
init sequence for when Kuryr is run in a Docker container. This commit
reverts that behavior of the offending patch, which gets Kuryr initialization
to work again and correctly detect the "tags" service plugin in Neutron.

It may not be the correct fix, but we can iterate on this one to find a
balance for Kuryr both inside and outside a Docker container.

Change-Id: I320f539f929be4955f401f363b23b7af09a7c73f
Signed-off-by: Kyle Mestery <mestery@mestery.com>