Do not attempt to re-install docker, regardless of its running status
Commit c96032ecc11b44b53f60c6240ffb05439fe57a91 changed the devstack
plugin to avoid reinstalling docker if the service was running.
However, there are a few meaningful cases where the service is not
running but the package is however installed and the user does not
want devstack to reinstall it.

This patch uses the distro package manager to find whether
docker-engine is insstalled already, rather than systemd or init.d.
The docker installation script will take care of cases where docker
was manually installed from binaries.

Change-Id: I6f66fdd3c35739486a6453b484eab81dfb4a9545