fix python packaging
After the split, it is necessary to make the metadata that is used to
generate the egg name reflect that this repo is housing kuryr-lib and
not kuryr. Also, since the versioning is handled by pbr by checking
tags, I removed that field so the egg has also the good version number.

Change-Id: Ic4cf2aea9f90c68de649f566e0c3d7cd6abaab2c
Signed-off-by: Antoni Segura Puimedon <antonisp@celebdor.com>