requirements: add missing oslo i18n dep in order
Somehow we were missing this dep. I suspect that another oslo dep was
pulling it anyway, but the policy is that any module that we import
directly should be in our requirements.txt even if it is being pulled
by another dependency (as that dependency could drop the sub dep and
we'd break).

Change-Id: I5d5ee420273d87c73c5c6c8ac806eaea2ce18c2b
Signed-off-by: Antoni Segura Puimedon <antonisp@celebdor.com>