Add randomness to the returned vlan_ids
This patch modifies allocate_segmentation_id so that the vlan_id
returned is randomly selected between the available ids.
The main rationale behind is to avoid vlan_id collisions, for
instance in active-active deployments, among different calls.
Note this will minimize the collision if nested containers are
created by using both kubernetes and docker in the same VM.

Partially Implements blueprint containers-in-instances

Change-Id: Iba2c94707c4932d59483c793452da88c43d4215c